application/requirements.txt
pywebview
uvicorn
jinja2
fastapi
pandas
tabulate
mysql-connector-python